create overview component for customer statistics
